requirements.txt
charms.reactive
charmhelpers
configparser
